* Move AVRCodeMapBase classes to parsers/code_map.py * Rename parsers/__init__.py to parsers/parse.py
